fixing metadata issue
